+++ lightningd/peer_control.c
@@ -57,22 +57,12 @@
 #include <lightningd/peer_htlcs.h>
 #include <lightningd/plugin_hook.h>
 #include <limits.h>
-#include <openingd/dualopend_wiregen.h>
 #include <stdlib.h>
 #include <unistd.h>
 #include <wally_bip32.h>
 #include <wire/common_wiregen.h>
 #include <wire/onion_wire.h>
 #include <wire/wire_sync.h>
-
-struct close_command {
-	/* Inside struct lightningd close_commands. */
-	struct list_node list;
-	/* Command structure. This is the parent of the close command. */
-	struct command *cmd;
-	/* Channel being closed. */
-	struct channel *channel;
-};
 
 static void destroy_peer(struct peer *peer)
 {
@@ -216,116 +206,6 @@ static void sign_last_tx(struct channel *channel,
 static void remove_sig(struct bitcoin_tx *signed_tx)
 {
 	bitcoin_tx_input_set_witness(signed_tx, 0, NULL);
-}
-
-/* Resolve a single close command. */
-static void
-resolve_one_close_command(struct close_command *cc, bool cooperative)
-{
-	struct json_stream *result = json_stream_success(cc->cmd);
-	struct bitcoin_txid txid;
-
-	bitcoin_txid(cc->channel->last_tx, &txid);
-
-	json_add_tx(result, "tx", cc->channel->last_tx);
-	json_add_txid(result, "txid", &txid);
-	if (cooperative)
-		json_add_string(result, "type", "mutual");
-	else
-		json_add_string(result, "type", "unilateral");
-
-	was_pending(command_success(cc->cmd, result));
-}
-
-/* Resolve a close command for a channel that will be closed soon. */
-static void
-resolve_close_command(struct lightningd *ld, struct channel *channel,
-		      bool cooperative)
-{
-	struct close_command *cc;
-	struct close_command *n;
-
-	list_for_each_safe (&ld->close_commands, cc, n, list) {
-		if (cc->channel != channel)
-			continue;
-		resolve_one_close_command(cc, cooperative);
-	}
-}
-
-/* Destroy the close command structure in reaction to the
- * channel being destroyed. */
-static void
-destroy_close_command_on_channel_destroy(struct channel *_ UNUSED,
-					 struct close_command *cc)
-{
-	/* The cc has the command as parent, so resolving the
-	 * command destroys the cc and triggers destroy_close_command.
-	 * Clear the cc->channel first so that we will not try to
-	 * remove a destructor. */
-	cc->channel = NULL;
-	was_pending(command_fail(cc->cmd, LIGHTNINGD,
-				 "Channel forgotten before proper close."));
-}
-
-/* Destroy the close command structure. */
-static void
-destroy_close_command(struct close_command *cc)
-{
-	list_del(&cc->list);
-	/* If destroy_close_command_on_channel_destroy was
-	 * triggered beforehand, it will have cleared
-	 * the channel field, preventing us from removing it
-	 * from an already-destroyed channel. */
-	if (!cc->channel)
-		return;
-	tal_del_destructor2(cc->channel,
-			    &destroy_close_command_on_channel_destroy,
-			    cc);
-}
-
-/* Handle timeout. */
-static void
-close_command_timeout(struct close_command *cc)
-{
-	/* This will trigger drop_to_chain, which will trigger
-	 * resolution of the command and destruction of the
-	 * close_command. */
-	json_notify_fmt(cc->cmd, LOG_INFORM,
-			"Timed out, forcing close.");
-	channel_fail_permanent(cc->channel, REASON_USER,
-			       "Forcibly closed by `close` command timeout");
-}
-
-/* Construct a close command structure and add to ld. */
-static void
-register_close_command(struct lightningd *ld,
-		       struct command *cmd,
-		       struct channel *channel,
-		       unsigned int timeout)
-{
-	struct close_command *cc;
-	assert(channel);
-
-	cc = tal(cmd, struct close_command);
-	list_add_tail(&ld->close_commands, &cc->list);
-	cc->cmd = cmd;
-	cc->channel = channel;
-	tal_add_destructor(cc, &destroy_close_command);
-	tal_add_destructor2(channel,
-			    &destroy_close_command_on_channel_destroy,
-			    cc);
-
-	if (!channel->owner) {
-		char *msg = tal_strdup(tmpctx, "peer is offline, will negotiate once they reconnect");
-		if (timeout)
-			tal_append_fmt(&msg, " (%u seconds before unilateral close)",
-				       timeout);
-		json_notify_fmt(cmd, LOG_INFORM, "%s.", msg);
-	}
-	log_debug(ld->log, "close_command: timeout = %u", timeout);
-	if (timeout)
-		new_reltimer(ld->timers, cc, time_from_sec(timeout),
-			     &close_command_timeout, cc);
 }
 
 static bool invalid_last_tx(const struct bitcoin_tx *tx)
@@ -1610,7 +1490,7 @@ static const struct json_command listpeers_command = {
 /* Comment added to satisfice AUTODATA */
 AUTODATA(json_command, &listpeers_command);
 
-static struct command_result *
+struct command_result *
 command_find_channel(struct command *cmd,
 		     const char *buffer, const jsmntok_t *tok,
 		     struct channel **channel)
@@ -1646,307 +1526,6 @@ command_find_channel(struct command *cmd,
 					     "should be a channel ID or short channel ID");
 	}
 }
-
-static struct command_result *param_outpoint(struct command *cmd,
-					     const char *name,
-					     const char *buffer,
-					     const jsmntok_t *tok,
-					     struct bitcoin_outpoint **outp)
-{
-	*outp = tal(cmd, struct bitcoin_outpoint);
-	if (json_to_outpoint(buffer, tok, *outp))
-		return NULL;
-	return command_fail_badparam(cmd, name, buffer, tok,
-				     "should be a txid:outnum");
-}
-
-static struct command_result *param_feerate_range(struct command *cmd,
-						  const char *name,
-						  const char *buffer,
-						  const jsmntok_t *tok,
-						  u32 **feerate_range)
-{
-	struct command_result *ret;
-	u32 *rate;
-
-	*feerate_range = tal_arr(cmd, u32, 2);
-	if (tok->type != JSMN_ARRAY || tok->size != 2)
-		return command_fail_badparam(cmd, name, buffer, tok,
-					     "should be an array of 2 entries");
-
-	ret = param_feerate(cmd, name, buffer, tok+1, &rate);
-	if (ret)
-		return ret;
-	(*feerate_range)[0] = *rate;
-	ret = param_feerate(cmd, name, buffer, tok+2, &rate);
-	if (ret)
-		return ret;
-	(*feerate_range)[1] = *rate;
-	return NULL;
-}
-
-static struct command_result *json_close(struct command *cmd,
-					 const char *buffer,
-					 const jsmntok_t *obj UNNEEDED,
-					 const jsmntok_t *params)
-{
-	const jsmntok_t *idtok;
-	struct peer *peer;
-	struct channel *channel COMPILER_WANTS_INIT("gcc 7.3.0 fails, 8.3 OK");
-	unsigned int *timeout;
-	const u8 *close_to_script = NULL;
-	bool close_script_set, wrong_funding_changed, *force_lease_close;
-	const char *fee_negotiation_step_str;
-	struct bitcoin_outpoint *wrong_funding;
-	u32 *feerate_range;
-	char* end;
-	bool anysegwit;
-
-	if (!param(cmd, buffer, params,
-		   p_req("id", param_tok, &idtok),
-		   p_opt_def("unilateraltimeout", param_number, &timeout,
-			     48 * 3600),
-		   p_opt("destination", param_bitcoin_address, &close_to_script),
-		   p_opt("fee_negotiation_step", param_string,
-			 &fee_negotiation_step_str),
-		   p_opt("wrong_funding", param_outpoint, &wrong_funding),
-		   p_opt_def("force_lease_closed", param_bool,
-			     &force_lease_close, false),
-		   p_opt("feerange", param_feerate_range, &feerate_range),
-		   NULL))
-		return command_param_failed();
-
-	peer = peer_from_json(cmd->ld, buffer, idtok);
-	if (peer)
-		channel = peer_active_channel(peer);
-	else {
-		struct command_result *res;
-		res = command_find_channel(cmd, buffer, idtok, &channel);
-		if (res)
-			return res;
-	}
-
-	if (!channel && peer) {
-		struct uncommitted_channel *uc = peer->uncommitted_channel;
-		if (uc) {
-			/* Easy case: peer can simply be forgotten. */
-			kill_uncommitted_channel(uc, "close command called");
-			goto discard_unopened;
-		}
-		if ((channel = peer_unsaved_channel(peer))) {
-			channel_unsaved_close_conn(channel,
-						   "close command called");
-			goto discard_unopened;
-		}
-		return command_fail(cmd, LIGHTNINGD,
-				    "Peer has no active channel");
-	}
-
-	if (!*force_lease_close && channel->opener != LOCAL
-	    && get_block_height(cmd->ld->topology) < channel->lease_expiry)
-		return command_fail(cmd, LIGHTNINGD,
-				    "Peer leased this channel from us, we"
-				    " shouldn't close until lease has expired"
-				    " (lease expires block %u,"
-				    " current block %u)",
-				    channel->lease_expiry,
-				    get_block_height(cmd->ld->topology));
-
-	/* If we've set a local shutdown script for this peer, and it's not the
-	 * default upfront script, try to close to a different channel.
-	 * Error is an operator error */
-	if (close_to_script && channel->shutdown_scriptpubkey[LOCAL]
-			&& !memeq(close_to_script,
-				  tal_count(close_to_script),
-				  channel->shutdown_scriptpubkey[LOCAL],
-				  tal_count(channel->shutdown_scriptpubkey[LOCAL]))) {
-		u8 *default_close_to = p2wpkh_for_keyidx(tmpctx, cmd->ld,
-							 channel->final_key_idx);
-		if (!memeq(default_close_to, tal_count(default_close_to),
-			   channel->shutdown_scriptpubkey[LOCAL],
-			   tal_count(channel->shutdown_scriptpubkey[LOCAL]))) {
-			return command_fail(cmd, JSONRPC2_INVALID_PARAMS,
-					    "Destination address %s does not match "
-					    "previous shutdown script %s",
-					    tal_hex(tmpctx, channel->shutdown_scriptpubkey[LOCAL]),
-					    tal_hex(tmpctx, close_to_script));
-		} else {
-			channel->shutdown_scriptpubkey[LOCAL] =
-				tal_free(channel->shutdown_scriptpubkey[LOCAL]);
-			channel->shutdown_scriptpubkey[LOCAL] =
-				tal_steal(channel, close_to_script);
-			close_script_set = true;
-		}
-	} else if (close_to_script && !channel->shutdown_scriptpubkey[LOCAL]) {
-		channel->shutdown_scriptpubkey[LOCAL]
-			= tal_steal(channel, cast_const(u8 *, close_to_script));
-		close_script_set = true;
-	} else if (!channel->shutdown_scriptpubkey[LOCAL]) {
-		channel->shutdown_scriptpubkey[LOCAL]
-			= p2wpkh_for_keyidx(channel, cmd->ld, channel->final_key_idx);
-		/* We don't save the default to disk */
-		close_script_set = false;
-	} else
-		close_script_set = false;
-
-	/* Don't send a scriptpubkey peer won't accept */
-	anysegwit = feature_negotiated(cmd->ld->our_features,
-				       channel->peer->their_features,
-				       OPT_SHUTDOWN_ANYSEGWIT);
-	if (!valid_shutdown_scriptpubkey(channel->shutdown_scriptpubkey[LOCAL],
-					 anysegwit)) {
-		/* Explicit check for future segwits. */
-		if (!anysegwit &&
-		    valid_shutdown_scriptpubkey(channel->shutdown_scriptpubkey
-						[LOCAL], true)) {
-			return command_fail(cmd, JSONRPC2_INVALID_PARAMS,
-					    "Peer does not allow v1+ shutdown addresses");
-		}
-
-		return command_fail(cmd, JSONRPC2_INVALID_PARAMS,
-				    "Invalid close destination");
-	}
-
-	if (fee_negotiation_step_str == NULL) {
-		channel->closing_fee_negotiation_step = 50;
-		channel->closing_fee_negotiation_step_unit =
-		    CLOSING_FEE_NEGOTIATION_STEP_UNIT_PERCENTAGE;
-	} else {
-		channel->closing_fee_negotiation_step =
-		    strtoull(fee_negotiation_step_str, &end, 10);
-
-		if (channel->closing_fee_negotiation_step == 0)
-			return command_fail(
-			    cmd, JSONRPC2_INVALID_PARAMS,
-			    "Wrong value given for fee_negotiation_step: "
-			    "\"%s\", must be positive",
-			    fee_negotiation_step_str);
-		else if (*end == '%') {
-			if (channel->closing_fee_negotiation_step > 100)
-				return command_fail(
-				    cmd, JSONRPC2_INVALID_PARAMS,
-				    "Wrong value given for "
-				    "fee_negotiation_step: \"%s\", the "
-				    "percentage should be between 1 and 100",
-				    fee_negotiation_step_str);
-			channel->closing_fee_negotiation_step_unit =
-			    CLOSING_FEE_NEGOTIATION_STEP_UNIT_PERCENTAGE;
-		} else if (*end == '\0')
-			channel->closing_fee_negotiation_step_unit =
-			    CLOSING_FEE_NEGOTIATION_STEP_UNIT_SATOSHI;
-		else
-			return command_fail(
-			    cmd, JSONRPC2_INVALID_PARAMS,
-			    "Wrong value given for fee_negotiation_step: "
-			    "\"%s\", should be an integer or an integer "
-			    "followed by %%",
-			    fee_negotiation_step_str);
-	}
-
-	if (wrong_funding) {
-		if (!feature_negotiated(cmd->ld->our_features,
-					channel->peer->their_features,
-					OPT_SHUTDOWN_WRONG_FUNDING)) {
-			return command_fail(cmd, JSONRPC2_INVALID_PARAMS,
-					    "wrong_funding feature not negotiated"
-					    " (we said %s, they said %s: try experimental-shutdown-wrong-funding?)",
-					    feature_offered(cmd->ld->our_features
-							    ->bits[INIT_FEATURE],
-							    OPT_SHUTDOWN_WRONG_FUNDING)
-					    ? "yes" : "no",
-					    feature_offered(channel->peer->their_features,
-							    OPT_SHUTDOWN_WRONG_FUNDING)
-					    ? "yes" : "no");
-		}
-
-		wrong_funding_changed = true;
-		channel->shutdown_wrong_funding
-			= tal_steal(channel, wrong_funding);
-	} else {
-		if (channel->shutdown_wrong_funding) {
-			channel->shutdown_wrong_funding
-				= tal_free(channel->shutdown_wrong_funding);
-			wrong_funding_changed = true;
-		} else
-			wrong_funding_changed = false;
-	}
-
-	/* May already be set by previous close cmd. */
-	tal_free(channel->closing_feerate_range);
-
-	/* Works fine if feerate_range is NULL */
-	channel->closing_feerate_range = tal_steal(channel, feerate_range);
-
-	/* Normal case.
-	 * We allow states shutting down and sigexchange; a previous
-	 * close command may have timed out, and this current command
-	 * will continue waiting for the effects of the previous
-	 * close command. */
-
-	/* If normal or locking in, transition to shutting down
-	 * state.
-	 * (if already shutting down or sigexchange, just keep
-	 * waiting) */
-	switch (channel->state) {
-		case CHANNELD_NORMAL:
-		case CHANNELD_AWAITING_LOCKIN:
-		case DUALOPEND_AWAITING_LOCKIN:
-			channel_set_state(channel,
-					  channel->state, CHANNELD_SHUTTING_DOWN,
-					  REASON_USER,
-					  "User or plugin invoked close command");
-			/* fallthrough */
-		case CHANNELD_SHUTTING_DOWN:
-			if (channel->owner) {
-				u8 *msg;
-				if (streq(channel->owner->name, "dualopend")) {
-					msg = towire_dualopend_send_shutdown(
-						NULL,
-						channel->shutdown_scriptpubkey[LOCAL]);
-				} else
-					msg = towire_channeld_send_shutdown(
-						NULL,
-						channel->shutdown_scriptpubkey[LOCAL],
-						channel->shutdown_wrong_funding);
-				subd_send_msg(channel->owner, take(msg));
-			}
-
-			break;
-		case CLOSINGD_SIGEXCHANGE:
-			break;
-		default:
-			return command_fail(cmd, LIGHTNINGD, "Channel is in state %s",
-					    channel_state_name(channel));
-	}
-
-	/* Register this command for later handling. */
-	register_close_command(cmd->ld, cmd, channel, *timeout);
-
-	/* If we set `channel->shutdown_scriptpubkey[LOCAL]` or
-	 * changed shutdown_wrong_funding, save it. */
-	if (close_script_set || wrong_funding_changed)
-		wallet_channel_save(cmd->ld->wallet, channel);
-
-	/* Wait until close drops down to chain. */
-	return command_still_pending(cmd);
-
-discard_unopened: {
-	struct json_stream *result = json_stream_success(cmd);
-	json_add_string(result, "type", "unopened");
-	return command_success(cmd, result);
-	}
-}
-
-static const struct json_command close_command = {
-	"close",
-	"channels",
-	json_close,
-	"Close the channel with {id} "
-	"(either peer ID, channel ID, or short channel ID). "
-	"Force a unilateral close after {unilateraltimeout} seconds (default 48h). "
-	"If {destination} address is provided, will be used as output address."
-};
-AUTODATA(json_command, &close_command);
 
 static void activate_peer(struct peer *peer, u32 delay)
 {
